Move DarkModeContext to contexts folder Rename DarkModeContext to ThemeModeContext
