--- kafka-ui-react-app/src/components/common/ThemeModeContext/index.ts
@@ -1,1 +1,0 @@
-export * from './ThemeModeContext';
